added utils/output_directory_getters.py
