+++ stage_scenes.py
@@ -2,12 +2,11 @@ import inspect
 import os
 import sys
 
-from constants import ANIMATIONS_DIR
 from constants import PRODUCTION_QUALITY_CAMERA_CONFIG
 from constants import PRODUCTION_QUALITY_FRAME_DURATION
 from extract_scene import get_module
 from extract_scene import is_scene
-from scene.scene import get_movie_output_directory
+from utils.output_directory_getters import get_movie_output_directory
 
 
 def get_sorted_scene_classes(module_name):
